requirements.txt
beautifulsoup4==4.7.1
paho-mqtt==1.4.0